refactor to extract the irs notice and case info into separate jsx components
